One more print(d) after For Line loop
If start marker present at beginning of each loop, AND a dict is not
empty, write the dict out. This get's the first to next to last records.
THEN after the loop closes, write the dict out, this gets the last
record. Not the most intuitive pattern, but seems to work.